Pull the online check into a different page from the data loading
It simplifies things by doing an automatic redirect when it goes online.
