+++ __tests__/features/accessibility.steps.ts
@@ -5,14 +5,14 @@ import Then from "../helpers/steps/Then";
 import When from "../helpers/steps/When";
 
 defineFeature(loadFeature("./accessibility.feature"), test => {
-  test("Index page is accessible", ({ defineStep, then }) => {
+  test("Loading page is accessible", ({ defineStep, then }) => {
     When.iVisitX(defineStep);
 
     Then.thePageShouldBeAccessible(defineStep);
 
     then("the page should have a descriptive title", async () => {
       await expect(browser!.getTitle()).resolves.toEqual(
-        "Tenancy and Household Check - Start"
+        "Tenancy and Household Check - Loading"
       );
     });
   });
